+++ src/main/java/stefan/licenta/homefinder/dto/AdDetailsDto.java
@@ -17,7 +17,7 @@ public class AdDetailsDto {
     private String title;
     private String adType;
     private String description;
-    private Integer price;
+    private Double price;
     private Integer surface;
     private Integer rooms;
     private String adItemType;
